requirements.txt
psutil
PyQt5
py-cpuinfo~=9.0.0